Moved Knn examples to Knn_algorithm class.
